requirements.txt
aioredis==1.3.1
amqp==2.6.1
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
auditlog3==1.0.1
autobahn==21.3.1
Automat==20.2.0
Babel==2.9.1
backcall==0.2.0
billiard==3.5.0.5
boto3==1.16.0
botocore==1.20.82
CacheControl==0.12.6
cachetools==4.2.2
celery==4.2.1
certifi==2020.12.5
cffi==1.14.5
channels==2.4.0
channels-redis==3.0.0
chardet==4.0.0
click==6.7
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.2
cryptography==3.4.7
daphne==2.5.0
decorator==5.1.0
deepmerge==0.1.0
Django==2.2
django-3-jet==1.0.8
django-autofixture==0.12.1
django-background-tasks==1.2.5
django-braces==1.14.0
django-celery-results==1.0.4
django-choices==1.7.1
django-ckeditor==6.3.0
django-coinpayments==0.1.3
django-compat==1.0.15
django-cors-headers==2.2.0
django-crontab==0.7.1
django-debug-toolbar==2.2
django-debug-toolbar-request-history==0.0.7
django-dotenv==1.4.2
django-excel-response==2.0.5
django-extra-fields==1.0.0
django-filter==2.0.0
django-jet==1.0.8
django-js-asset==2.0.0
django-mathfilters==1.0.0
django-model-utils==4.2.0
django-nose==1.4.6
django-picklefield==2.0
django-redis==4.12.1
django-rest-swagger==2.2.0
django-storages==1.7.1
django-user-agents==0.4.0
django-utils-six==2.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.0.0
drf-yasg==1.10.2
et-xmlfile==1.1.0
factory-boy==2.11.1
Faker==8.3.0
firebase-admin==4.3.0
flex==6.13.2
flower==0.9.3
google-api-core==1.28.0
google-api-python-client==2.6.0
google-auth==1.30.1
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-firestore==2.1.1
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
grpcio==1.38.0
gunicorn==20.1.0
h11==0.12.0
hiredis==2.0.0
httplib2==0.19.1
humanize==3.2.0
hyperlink==21.0.0
html2text==2020.1.16
idna==2.10
incremental==21.3.0
inflection==0.5.1
ipython==7.30.0
itypes==1.2.0
jdcal==1.4.1
jedi==0.18.1
Jinja2==3.0.1
jmespath==0.10.0
jsondiff==1.2.0
jsonmerge==1.7.0
jsonpointer==1.14
jsonschema==3.2.0
kombu==4.6.11
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mercadopago==2.0.6
msgpack==1.0.2
newrelic==5.22.1.152
nose==1.3.7
numpy==1.21.5
openapi-codec==1.3.2
openpyxl==2.4.11
packaging==20.9
pandas==1.3.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pika==1.1.0
Pillow==9.1.1
prometheus-client==0.8.0
prompt-toolkit==3.0.16
proto-plus==1.18.1
protobuf==3.17.1
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
Pygments==2.10.0
PyJWT==1.7.1
pynamodb==4.3.2
pyOpenssL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-crontab==2.3.5
python-dateutil==2.8.1
python-http-client==3.3.4
pytz==2021.1
PyYAML==3.11
redis==3.5.0
requests==2.25.1
requests-aws4auth==1.1.0
rfc3987==1.3.8
rollbar==0.15.2
rsa==4.7.2
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
s3transfer==0.3.7
sendgrid==6.9.7
service-identity==21.1.0
simplejson==3.17.2
six==1.16.0
sqlparse==0.4.1
starkbank-ecdsa==2.0.3
strict-rfc3339==0.7
swagger-spec-validator==2.3.1
text-unidecode==1.3
tornado==5.1.1
traitlets==5.1.1
twilio==6.63.0
Twisted==21.2.0
txaio==21.2.1
ua-parser==0.10.0
uritemplate==3.0.1
urllib3==1.26.5
user-agents==2.2.0
validate-email==1.3
vine==1.3.0
wcwidth==0.2.5
xmltodict==0.12.0
zope.interface==5.4.0
xhtml2pdf==0.2.9
pyotp==2.8.0
django-tinymce==3.6.1
